+++ CHANGELOG.md
@@ -17,6 +17,7 @@ recommendations of [keepachangelog.com](http://keepachangelog.com/).
 - Add webhook management for Post creation, updates, deletion and publication
 - Updated Suggest Tags input to accept freeInput option (defaults to true) - when set to false only provided typeahead tags are valid
 - Set default title and notifications to Integral Backend CRUD endpoints
+- Pages - Added 'Archived' status
 
 ### Fixed